+++ rxlifecycle-android-lifecycle-kotlin/build.gradle
@@ -18,8 +18,9 @@ repositories {
 
 dependencies {
     compile kotlinStdlib
-    compile project(':rxlifecycle-android')
+    compile project(':rxlifecycle-android-lifecycle')
 }
 
 apply from: "$rootDir/gradle/artifacts.gradle"
 apply from: "$rootDir/gradle/gradle-mvn-push.gradle"
+
