+++ rxlifecycle-android-lifecycle-kotlin/gradle.properties
@@ -1,0 +1,4 @@
+POM_NAME=RxLifecycle-Andoird-Lifecycle-kotlin
+POM_DESCRIPTION=RxLifecycle-Android-Lifecycle-kotlin
+POM_ARTIFACT_ID=rxlifecycle-android-lifecycle-kotlin
+POM_PACKAGING=aar
