+++ rxlifecycle-android-lifecycle-kotlin/src/main/AndroidManifest.xml
@@ -1,0 +1,2 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="me.tatarka.rxlifecycle_android_lifecycle_kotlin" />
