+++ rxlifecycle-android-lifecycle/src/test/java/com/trello/rxlifecycle2/android/lifecycle/RxLifecycleObserverActivityTest.java
@@ -5,8 +5,8 @@ import android.arch.lifecycle.LifecycleActivity;
 import android.arch.lifecycle.LifecycleOwner;
 import android.arch.lifecycle.LifecycleRuntimeTrojanProvider;
 
+import com.trello.lifecycle2.android.lifecycle.RxLifecycleObserver;
 import com.trello.rxlifecycle2.LifecycleProvider;
-import com.trellow.relifecycle2.android.lifecycle.RxLifecycleObserver;
 
 import org.junit.Before;
 import org.junit.Test;
